feat: Descriptor adapters
Created a descriptor adapter to be used to write only the nodes of the descripters defined in the schema config.
This adapter is completely hardcoded and might be changed in the future.
The adapter also don't contain any function to write edges as that will done in the compound adapter